test: Add helper library to fake passwd/group files
The requirements for testing are currently that the system have several
users and groups predefined.  The chosen users and groups are typically
found on every system so that's a safe bet.  However, tests for the quote
code may involve more esoteric names that include backslashes or other
quoted characters.  This patch adds a helper library that implements
the passwd and group calls and redirects the reads to project-defined
passwd and group files.  That way, we know for certain that those
usernames and groups will be found during testing.  The helper library is
enabled using LD_PRELOAD via a wrapper script.  The library will not be
installed as part of the project's make install.  Since the local user
might not be found in the local test.{group,passwd}, we extend test/run
to use numeric uid/gids if the lookup fails.