quote: escape literal backslashes
The octal unquote code can't handle escaping backslashes except by
specifying the octal code for the backslash.  That's not exactly a
user-friendly interface for a scenario that may not be that uncommon.
The unquote code isn't documented in the man page and can result in
confusion when specifying names via a Windows domain that are entirely
numeric (e.g. WINDOM\1234).

This patch adds handling to allow literal escaping of the backslash (\\)
as both input and output and documents the quoting rules in the setfacl
man page.  Also included are test cases for escaped backslashes followed
by letters, escaped backslashes followed by numbers, and the original
use case of escaped character literals.

[Minor fixes by Andreas Gruenbacher.]