build: Fix cross-compilation on musl-based systems
The gettext implementation in musl does not define some of the internal
symbols that AM_GNU_GETTEXT used to check for. This got fixed in gettext
0.19.8, but the AM_GNU_GETTEXT_VERSION macro requires a specific version
even if the system contains a higher one. We switch to use the new
AM_GNU_GETTEXT_REQUIRE_VERSION macro which requires at least a specific
version, and bump the AM_GNU_GETTEXT_VERSION to the first one that had
support for the new macro.

Signed-off-by: Guillem Jover <guillem@hadrons.org>