licensing: clean up LGPL boilerplate for library code
The library headers & sources are intended to be LGPL-2.1.  A few
files are confused and say both GPL-2 & LGPL-2.1, and a few say
they are only GPL-2.  Fix the headers for all library files to be
only LGPL-2.1.

While we're here, harmonize the boilerplate across all of them as
even though most are already LGPL-2.1, they've formatted the header
text in slightly different ways.

A few files under tools/ that are licensed as GPL-2 are left as-is.
It's not clear as to the intentions there, so leave it up to the
original author to decide.

Bug: https://savannah.nongnu.org/bugs/index.php?64284