libmisc: internalize walk_tree API
While shared libs take care of hiding these symbols from the exported
list, they can still show up in static linking and cause collisions.
The codebase is using __acl_ as a prefix for other internal symbols,
so rename these to match.

We can cheat here with a define so the rest of the codebase doesn't
need to be changed to the more verbose form.