acl_copy_entry: Prevent accidental NULL pointer dereference
In acl_copy_entry(), when dest_d turns out to be invalid, dest_p will be
NULL.  Instead of checking for that, we are accidentally checking if
dest_d is NULL.  As a result, when called with an invalid dest_d object,
acl_copy_entry() will cause a NULL pointer dereference instead of
indicating an error.  This is a relatively minor problem, but worth
fixing nonetheless.

Reported-by: His Shadow <shadowpilot34@gmail.com>
Signed-off-by: Andreas Gruenbacher <agruenba@redhat.com>